+++ app/Http/Controllers/OrderDetailController.php
@@ -17,8 +17,15 @@ class OrderDetailController extends Controller
      */
     public function index()
     {
-        $orderdetails = OrderDetail::with('order','product')->paginate(3);
-        return view('admin.orderdetail.index',compact(['orderdetails']));
+        try {
+            $this->authorize('viewAny',OrderDetail::class);
+            $orderdetails = OrderDetail::with('order','product')->paginate(3);
+            return view('admin.orderdetail.index',compact(['orderdetails']));
+            //code...
+        } catch (\Exception $e) {
+            alert()->warning('Have problem! Please try again late');
+            return back();
+        } 
     }
 
     /**
@@ -26,8 +33,15 @@ class OrderDetailController extends Controller
      */
     public function create(String $order_id)
     {
-        $products = Product::get();
-        return view('admin.orderdetail.create',compact(['products','order_id']));
+        try {
+            //code...
+            $this->authorize('create',OrderDetail::class);
+            $products = Product::get();
+            return view('admin.orderdetail.create',compact(['products','order_id']));
+        } catch (\Exception $e) {
+            alert()->warning('Have problem! Please try again late');
+            return back();
+        } 
     }
 
     /**
@@ -67,10 +81,18 @@ class OrderDetailController extends Controller
      */
     public function edit(String $id)
     {
-        $detail = OrderDetail::find($id);
-        $products = Product::get();
-        $orders = Order::get();
-        return view('admin.orderdetail.edit',compact(['detail','products','orders']));
+        try {
+            //code...
+            $detail = OrderDetail::findOrFail($id);
+            $this->authorize('update', $detail);
+            $detail = OrderDetail::with('product')->find($id);
+            $products = Product::get();
+            $orders = Order::get();
+            return view('admin.orderdetail.edit',compact(['detail','products','orders']));
+        } catch (\Exception $e) {
+            alert()->warning('Have problem! Please try again late');
+            return back();
+        } 
     }
 
     /**
@@ -103,22 +125,41 @@ class OrderDetailController extends Controller
      */
     public function trash()
     {
-        $orderdetails = OrderDetail::onlyTrashed()->with('order','product')->paginate(3);
-        return view('admin.orderdetail.trash',compact(['orderdetails']));
+        try {
+            //code...
+            $this->authorize('viewTrash',OrderDetail::class);
+            $orderdetails = OrderDetail::onlyTrashed()->with('order','product')->paginate(3);
+            return view('admin.orderdetail.trash',compact(['orderdetails']));
+        } catch (\Exception $e) {
+            alert()->warning('Have problem! Please try again late');
+            return back();
+        } 
     }
     public function destroy(String $id)
     {
-        $orderdetail = OrderDetail::find($id);
-        $orderdetail->delete();
-        alert()->success('Success move to trash');
-        return back();
+        try {
+            //code...
+            $detail = OrderDetail::findOrFail($id);
+            $this->authorize('delete', $detail);
+            $orderdetail = OrderDetail::find($id);
+            $orderdetail->delete();
+            alert()->success('Success move to trash');
+            return back();
+        } catch (\Exception $e) {
+            alert()->warning('Have problem! Please try again late');
+            return back();
+        } 
     }
     function restore(String $id){
         try {
+            //code..
+            $order = Order::findOrFail($id);
             $orderdetail = OrderDetail::withTrashed()->find($id);
+            // dd($orderdetail->order_id);
+            $this->authorize('restore',$orderdetail);
             $orderdetail->restore();
             alert()->success('Restore order detail success');
-            return redirect()->route('order.show',$id);
+            return redirect()->route('order.show',$orderdetail->order_id);
         } catch (\Exception $e) {
             // Log::error($e->getMessage());
             alert()->warning('Have problem! Please try again late');
@@ -127,7 +168,9 @@ class OrderDetailController extends Controller
     }
     function deleteforever(String $id){
         try {
+            //code..
             $orderdetail = OrderDetail::withTrashed()->find($id);
+            $this->authorize('forceDelete',$orderdetail);
             $orderdetail->forceDelete();
             alert()->success('Destroy order success');
             return back();
